comleted exercise 6.3 storing votes in redux store.
