+++ part6/redux-anecdotes/src/App.js
@@ -1,13 +1,35 @@
-import NewAnecdote from './components/NewAnecdote'
-import Votes from './components/Votes'
+import { useSelector, useDispatch } from 'react-redux'
 
 const App = () => {
-  // * const importantNotes = useSelector(state => state.filter(note => note.important))
+  const anecdotes = useSelector((state) => state)
+  const dispatch = useDispatch()
+
+  const addVote = (id) => {
+    return {
+      type: 'VOTE',
+      payload: { id },
+    }
+  }
 
   return (
-    <div>
-      <NewAnecdote />
-      <Votes />
+    <div id='form'>
+      <h2>Anecdotes</h2>
+      {anecdotes.map((anecdote) => (
+        <div key={anecdote.id}>
+          <div id='list'>{anecdote.content}</div>
+          <div>
+            has {anecdote.votes}
+            <button onClick={() => dispatch(addVote(anecdote.id))}>vote</button>
+          </div>
+        </div>
+      ))}
+      <h2>create new</h2>
+      <form>
+        <div>
+          <input />
+        </div>
+        <button>create</button>
+      </form>
     </div>
   )
 }
