+++ part6/redux-anecdotes/src/app.css
@@ -3,7 +3,7 @@
 }
 
 #vote {
-  align-content: center;
+  /* align-content: center; */
   margin-left: 1rem;
 }
 button {
@@ -11,7 +11,7 @@ button {
   margin-left: 1rem;
 }
 #form {
-  text-align: center;
+  /* text-align: center; */
   margin-bottom: 2rem;
 }
 #list {
